refactor : Lottos.java
